requirements.txt
discord.py==1.7.1
isbnlib==3.10.6
mysql-connector-python==8.0.23
python-dotenv==0.17.0
